:truck: refactor: projects-overview page
- renames work-examples to projects overview
- rename all related files and links. Work examples was confusing
(specifically with project page) because the link is "Projects", the
page heading is "Projects" but it's actually the work example page.
- because the project grid used on the homepage and project overview
page are identical, it made the most sense to extract the duplicate code
into its own component for each template to utilize.
- markup for the project overview page uses BEM class naming
- include text content in MD data
- remove any duplicate code (most of the CSS in this file was duplicated
from the project-grid and/or the project-section partials)
- refactor SCSS and use nesting