:fire: refactor: move resume modifier to contact-section
Because the resume page is the only interior page that isn't the teal
background color, the overlay needs to be inverted for teh contact
section (with black instead of white). It made more sense to keep use a
cmp-contact--resume class in the contact-section.scss file, rather than
have a cmp-body--resume class in the resume-page.scss file.